fix: OpenClaw hook structure + revert package name to witness-memory-chain
- Restructure hooks to match OpenClaw's discovery format:
  - HOOK.md + handler.ts per hook directory
  - memory-chain-bootstrap for agent:bootstrap events
  - memory-chain-reset for command:reset events
- Revert package name from @witness/memory-chain to witness-memory-chain
  (npm org doesn't exist)
- Update all skill imports accordingly

The hooks now follow OpenClaw's expected structure and will be
properly discovered by the gateway.
